Shortened file names due to C1083 build error
