+++ CMakeLists.txt
@@ -19,13 +19,13 @@ set(ADP_SOURCES
 	src/codechandler/ttml/TTML.cpp
 	src/common/AdaptiveStream.cpp
 	src/common/AdaptiveTree.cpp
-	src/common/RepresentationChooser.cpp
-	src/common/RepresentationChooserAskQuality.cpp
-	src/common/RepresentationChooserDefault.cpp
-	src/common/RepresentationChooserFixedRes.cpp
-	src/common/RepresentationChooserManualOSD.cpp
-	src/common/RepresentationChooserTest.cpp
-	src/common/RepresentationSelector.cpp
+	src/common/Chooser.cpp
+	src/common/ChooserAskQuality.cpp
+	src/common/ChooserDefault.cpp
+	src/common/ChooserFixedRes.cpp
+	src/common/ChooserManualOSD.cpp
+	src/common/ChooserTest.cpp
+	src/common/ReprSelector.cpp
 	src/parser/DASHTree.cpp
 	src/parser/HLSTree.cpp
 	src/parser/SmoothTree.cpp
@@ -65,13 +65,13 @@ set(ADP_HEADERS
     src/common/AdaptiveDecrypter.h
 	src/common/AdaptiveStream.h
 	src/common/AdaptiveTree.h
-	src/common/RepresentationChooser.h
-	src/common/RepresentationChooserAskQuality.h
-	src/common/RepresentationChooserDefault.h
-	src/common/RepresentationChooserFixedRes.h
-	src/common/RepresentationChooserManualOSD.h
-	src/common/RepresentationChooserTest.h
-	src/common/RepresentationSelector.h
+	src/common/Chooser.h
+	src/common/ChooserAskQuality.h
+	src/common/ChooserDefault.h
+	src/common/ChooserFixedRes.h
+	src/common/ChooserManualOSD.h
+	src/common/ChooserTest.h
+	src/common/ReprSelector.h
 	src/parser/DASHTree.h
 	src/parser/HLSTree.h
 	src/parser/SmoothTree.h
